@@ -16,12 +16,21 @@ public class SpectatorBotCommand implements CommandExecutor {
 
   @Override
   public boolean onCommand(CommandSender sender, Command command, String label, String[] args) {
-
-    if (sender instanceof Player) {
-      Player player = (Player) sender;
-
+    System.out.println("spectatorbot command received");
+    if (sender instanceof Player player) {
+      System.out.println("Testing for permissions");
       if (player.isOp() || player.hasPermission("spectatorbot")) {
-
+        System.out.println("");
+        // If spectateManager is already using sender as a bot, turn off spectate mode
+        if (tphmc.spectateManager.getSpectateBot() != null && tphmc.spectateManager.getSpectateBot().getUniqueId().equals(player.getUniqueId())) {
+          System.out.println("Bot is no longer spectating");
+          tphmc.spectateManager.StopSpectate();
+        }
+        else {
+          System.out.println("Bot is starting to spectate");
+          tphmc.spectateManager.StartSpectate(player);
+        }
+        return true;
       }
     }
     return false;
